Ram is now displayed on the lcd.

+++ G19Tastatur Applet/hardware.cpp
@@ -172,5 +172,5 @@ unsigned short Hardware::getVirtualMemoryCurrentlyUsed() {
 	MEMORYSTATUSEX statex;
 	statex.dwLength = sizeof(statex);
 	GlobalMemoryStatusEx(&statex);
-	return statex.dwMemoryLoad;
+	return static_cast<unsigned short>(statex.dwMemoryLoad);
 }
